move powf_scalar and unary_simd
